Rewrite with ebitengine instead of sdl+gtk (#66)
* vendor: add ebitengine, bump kong

* Switch to ebitengine (no audio nor input for now)

* Simplify cli

* Add missing cli.go

* Remove ui package

* vendor: remove gotk3 and sdl_ttf

* Use ebitengin for input (keyboard)

* Move output into emu package. Simplify headless emulation

* emu/rpc: remove package

* Rename package: s/ebit/ui

* vendor: add ebitenui

* ui: scale emu screen based on screen size

* vendor: gofont

* ui: "start" button starts emulator

* some progress

* ui: pause actually pauses emulation

* Update README.md

* ci: update dependencies for both workflows

* Add -v, --verbose cli flag to set global level from warn to info.

* Improve --help

* vendor: add ebitenutil

* ui: wip (race condition)

* ui+emu: move ui-related concerns from emu to ui

* ui: draw paused overlay

* ui: wip

* still race

* hw/shaders: remove

* hw/hwio: cosmetics

* hw: rewrite ppu

* hw: PPU nestest passes

* hw: pass dmc_dma/double_2007_read test

* hw: ppu passes sprite overflow tests

* hw/snapshot: regenerate PPU state

* hw: ppu renamings

* hw: pass vram_access.nes (palette read that still overwrite read buf)

* hw: fix dma_2007_read test

* emu/log: use local level for Disable

* ci: ebitengine requires $DISPLAY

* hw: ppu renaming

* hw: save/restore ppu state

* ui: poc ui with pause (reset and stop), rom_list and config states

* hw/apu: switch logs from info to debug

* CLI cosmetics

* hw/apu: cosmetics

* vendor: use ebitengine/oto/v3 for audio

* ui: use oto to render emulator audio

* ui: simplify running state, oto context creation.

* emu: improve concurrency in Emulator controls

* ui: nicer pause buttons

* Add fullscreen: toggle with F, and start_fullscreen config option

* ui: show FPS on title bar

* config: s/ConfigDir/Dir

* ui: recent roms [WIP]

* vendor: add statsviz

* Add -monitor option, enabling Statsviz on a given host:port

* Print stack trace on SIGUSR2 (unix-only)

* ui: improve roms list

* ui: fix click on Stop when paused

* emu: simplify Emulator loop

* ui: polish font loading

* Move logo to resource

* ui: menu exit

* ui: rom list layout, correctly stretching

* config: API cosmetics

* main cosmetics

* ui: cosmetics

* ui: add menu on recent_roms window

* ui: remove unused configState for now

* ui: cleanup unused stuff

* ui: rom_list: frameImage

* ui: rom_list: prepare for adding rom name label

* ui: add rom cell label

* ui: recent_roms: prepare for keyboard rom selection

* ui: select rom with arrow keys

* ui: reorganize rom list

* ui: nicer rom list

* ui: basic scrolling in rom list

* ui: scroll cleanup

* rom_list: scrolling works

* ui: rom_list cleanup

* Fix data race in sampleBuffer

* ui: handle window/screen size in application

* ui: cosmetics

* ui: simplify application initialization

* ui: only keep a single ebitenui pointer

* vendor: bump ebiten to v2.8.8

* ui: add assets

* vendor: add golang.org/x/text

* ui: compiles

* ui: config menu stub

* ui: fix resume didn't restart audio

* ui: split running and paused states

* ui: simplify paused state

* ui: homogenization of widgets style

* ui: simplify pause state

* emu: fix potential deadlock in pause/resume

* ui: config: 'Back to main menu' button

* ui: reorganize code

* vendor: bump ebitenui to v0.7.2

* ui: wip input config

* ui: add combo and list helpers

* ui: add empty video config page

* ui: set SingleThread:false to allow Ctrl-C to exit (cleanly)

* ui: use ptrTo rather than constantutil.ConstantToPointer

* ui: pass nestor config to configuraiton pages

* ui: load/save paddle preset

* ui: cleanup Combobox code

* ui: input config layout (AI)

* ui: add AGENTS.md

* ui: presetWrapper centered

* ui: Combobox embeds widget.ListComboButton

* ui: config page, align button assignment table

* ui: static table for paddle assignments

* ui: polish button input table

* ui: relayout experiment

* ui: working input table (UI only)

* ui: input capture stub (app state)

* hw: start input provider refactor

* Rework and simplify input configuration

* Input capture for paddle configuration (keyboard)

* vendor: remove SDL

* Paddle input configure & capture

* hw/input: export UnsetKey

* ui: pressing ESC during pad capture unsets the current mapping

* hw/input: don't mix input types for a given preset

* ci: start xvfb to fix glfw error at ebitengine init

* ui: polish input page

* ui: simplify input capture code

* hw/input: s/PadButton/Joystick

* hw/input+ui: use configured input for controls

* ui: add vsync and fullscreen to video config page

* go.mod: bump to go1.25.3

* ui: add newCheckbox2States helper

* ui: cleanup app initialization

* ui: start on monitor from config

* ui: add 'monitor' to config/video page

* ui: lint + cosmetics

* ui: add config/emulation page

* vendor: bump ebitengine, msgp and x/image

* vendor: add sqweek/dialog

* ui: add open ROM menu

* go generate

* ui: s/rom_list/main

* ui: Start ROM after file menu. Add to recent roms list (wip)

* ci: add libgtk-3-dev for sqweek/dialog

* ui: add error window

* ui: apply fragment shader on frame

* ui/shaders: add Names(). Add test to load all declared shaders

* ui/shaders: add passthrough and gizmo-crt shaders

* ui/shader: rename package s/shaders/shader

* ui: fix: get back to current config tag

* ui: fix typo in "run-ahead frame" combo box

* ui: read shader from config, add combobox to chose the shader

* ui/shader: add "scanline" shader

* ui/shader: add crt-nes-mini

* ui/shaders: add 3 new CRT shaders

* vendor: add ebitengine/debugui

* ui: allow changing shader at runtime (with debugui)

* vendor: bump statsviz to 0.8.0

* ui/shader: add 3 new shaders

* ui: set runnable on unfocused

This also enables stopping the application with Ctrl-C in the terminal

* ui/shader: rename shader extension to .kage

* ui/shader: homogeneize kage file header

* ui: cosmetics

* ui: move shader from app to runningState

* ui: simplify F11 handling (fullscreen on/off)

* ui: F5 to show shader ui

+++ hw/tracer.go
@@ -14,7 +14,7 @@ type cpuState struct {
 
 	Clock    int64
 	PPUCycle uint32
-	Scanline int
+	Scanline int16
 }
 
 type disasmer interface {
